requirements.txt
Flask
python-dotenv
gigachat
Flask-SQLAlchemy
psycopg2-binary